Temoprary commit. The features are not completed yet.
Remove treeViewSet, listViewSet, etc (scrum task #1333)
Remove __set and __get from the index controller (task #1335)

Do some __call overwrite magic in the default model to avoid errors, as the default model cannot be a active record. We have to discuss, if we want to allow non ActiveRecord models, and if we have to define and interface for them.