NOTE: This is a temporary commit and contains changes that breaks all the code checkings and unit testings as well as the application itself at the moment.
Remove project specific code from the default indexcontroller and move it into the project index controller.
Rearrange assignments of template variables and remove __set and __get.
Remove paging code as it considered as not longer necessary. We want to do this later using ajax, json and js.
Implement an iterator interface into the activerecord pattern to make is possible to easy iterate through the actual columns (without the references of other tables).
The generation of an title array for the rendering of the listview is now done using a smary_modifier @titles (e.g. $records|@titles).
Refactor the database manager and add constants instead of passing various, non-documented strings to the methods. This simplifies reading. The db manager class is not initialized and configured with an active record.
Start refactoring of the form view, but at the moment this is broken.
Remove list.tpl, form.tpl, etc from the todo model as an example that it should be possible to render this with default views and it must not be necessary to have templates in an non-default module.