Some smaller improvements in code.
Don't access the _data array directly but use the mapping (e.g. $foo->_data['foo'] can be accessed using $foo->foo). Thats how ActiveRecord works and should be used.

Remove some unused methods.

If we return an ID and not an object itself we call the return function getFooId().
If a method is called getFoo() it's assumed to return an object.