Use BIGINT instead of INT for CRC32
CRC32 values are 32bit long. On 32bit machines PHPs integers as well as
mysql INT columns are 32 bit signed. On x64 the PHPs integers are 64bits long
and therefore the 32bit value fits into the 64bits integer as positive integers
instead of being negatives integers on 32 machines. But the mysql INT column is
32 bit and therefore we tried to add values that are too big for the INT column.
To prevend this we now use BIGINT which can handle the CRC32 integers on 64bit machines.