Make modifiers independent from the database layer
The modifiers and therefore the complete rendering of forms fails if the
model is not a database model. We change the instanceof checking and
add getRights to the global interface which every model must fit.

Also ehance the model information interface and add a dumb default model information
class that wraps array. This makes dealing with models, not connected with the database
e.g. AdminModels more easier.