Change the way the tree is outputed.
The tree is now part of the index.phtml to reduce the amount of templates.
In addition to that the tree is now rendered in a recursive way which is much
more intuitive than the linear approach with str_repeats. We now get a static but
beautiful tree. As we want to replace the tree component with the dojo tree, we don't
make any additional efford to make  toggled nodes