We separate jsonListAction into two different methods.
jsonListAction had a different behaviour depending on the view
parameter that was added. To make things cleaner and to avoid
undocumented return values due to some magic get parameters we
instead use jsonListAction to display lists and jsonTreeAction
to display trees.

Both are still prefixed with json as this will implicitly document
that we return json data.