By default the ViewRenderer helper should NOT render.
This must be done in the IndexController. When doing this in the Front Controller
configuration (htdocs/index.php) it will be overrided by the Zend_Action
which the IndexController must inherit. Therefore the only point to set this while
not disabling the ViewRenderer helper at all is the init method of
the IndexController