Forward PublishedExceptions to the frontend.
Thrown exception of type Phprojekt_PublishedExceptions are now forwarded
as JSON to the frontend. Exceptions that don't implement PublishedException
are not forwarded.

We might want to provide XML instead of JSON in the future. Then we need
a mechanism to do switch between requested formats. At the moment we only
do JSON and refactoring shouldn't be that hard.