Don't allow to throw exception as this will cause the ErrorController not to be called
The ErrorController logs our exceptions as well as distinguish between exceptions that
can be forwarded.
If Zend_Front_Controller::throwException is true, the ErrorController will not be called,
but instead no exception will be catched, leading to Fatal Errors. As a result, the
frontend never get's a proper exception but instead Fatal Errors.