New Mail Notification class (all the files involved). Note: copy and paste the new section inside configuration.ini-dist into your configuration.ini file ("; Mail Notification").
Summary: a new tab will be seen in the Project, Todo, Calendar and Note module forms (and also in future module forms).
Inside that tab, if the checkbox is on, an email will be sent to all the users with read access to the item, telling them about the creation or modification of that item, after pressing the button Save.