#hash implementation for the Default module
All the action changed from "reload", "openForm", "showTagResult", and "showSearchResult" to "setUrlHash"
The function return the #hash that is processed then by the processUrlHash function.

Also change the load and init functions for allow the processUrlHash in the start (Direct link).

The Direct link, change link (navigation and manual change) and Back button, work in FF
The Direct link, change link (navigation and manual change) work in IE (the Back button don't work)