* Bugs: CALENDAR - Multiple events (second commit)
Details:
The jsonGetParticipants method has been changed by jsonGetRelatedEvents, now it returns the participants for an event plus the related events and it is cached.
Now after each saving an event, the 3 urls for the saved event and related ones are deleted from cache (url, relatedDataUrl and tagUrl (previously only the main url).
Deleted submitForm function from Form.js and it is used the default one.
Added tests.