* Notification: improvement, new structure and simplification
1) Now Phprojekt_Notification and Phprojekt_Notification_Mail are independent, the first one doesn't extend anything and the second one extends Zend_Mail.
Phprojekt_Notification is called by Default_Helpers_Save with the selected transport as a parameter (currently MAIL_TEXT or MAIL_HTML, in the future maybe SMS, PHONE_CALL ;-), whatever is created as Phprojekt_Notification_<transport> ). Then Phprojekt_Notification calls the appropriate transport to send the notification, and provides the essential data for it (sender, recipients, body, etc). The purpose of this is that in the future it may be added easily new transports for the Notification (e.g.: sms).

2) Also general improvement and simplification of classes and methods.