Remove filesChanged, update the field always.
The filesChanged is a var that indicates if the field
should be updated or not with the current value of the files.
The field is always updated except on the fileFormAction.
That is not necessary, since in this action the field should be updated too,
so without the var and without the check,
now the field is updated always on each action.

Signed-off-by: Gustavo Solt <gustavo.solt@mayflower.de>