Move fileAction functions to a Helper.
For get less code in the IndexController,
all the upload actions are moved into a helper class,
and only the actions are in the controller.
This is the same structure for the save, delete and rights functions.
Then is easy to extend the upload action in the module controller to use other helper.

Signed-off-by: Gustavo Solt <gustavo.solt@mayflower.de>