Overwrite some functions to make sure our participantData is ours.
Because ActiveRecord doesn't return a new object on a call to find(), we
have to reset our participant cache in the php object, or the found
object will end up with our original participants.
Also, cloning an ActiveRecord object get's you a blank object, so we
also have to delete the participant data.