Add confirmation status changes.
Because the user interface doesn't make a difference between creating a
new element and editing an existing one, we show the confirmation status
field when creating events, too.

Instead of locking, we'll just allow different confirmation statuses for
the event owner.

We also need to return null instead of throwing an Exception in
Calendar2_Models_Calendar2->getConfirmationStatus() if the given user
does not participate in the event because it will get called on an empty
object when the user starts to fill in a new event.