Added FREQINTERVAL property to Calendar2_Helper_Rrule
Until now, the calendar used the FREQ-part of the recurrence rule to get the
interval between single occurrences. (Freq is something like daily, monthly
etc.) This is fine, as long as INTERVAL is 1 or not given. INTERVAL was not used
yet.
To tackle this, when parsing the rrule, a new value FREQINTERVAL is generated.
This is a DateInterval consisting of INTERVAL times the original interval as
extracted from the FREQ property.