Removed support for COUNT from Calendar2_Helper_Rrule
It was never really supported anyways, and instead of creating the illusion that
it is I decided to remove it.

Note that there is also a bug in the chunk after FREQINTERVAL is set.
The check shouldn't be whether both are null but whether both are set, as rrules
with both UNTIL and COUNT aren't allowed by rfc5545.