Fixed timezone bug in fetchAllForPeriod and findOccurrence.
It was not taken into account that activeRecord automatically converts
timestamps to usertime when __set is used.
Because of this, the user saw UTC times and jsonDetailAction didn't work
(because the client sent a wrong occurrence time).

This is solved like in other places, we just convert the utc times to usertimes
before assigning variables.