Made Calendar2_Models_Calendar2->_excludeDate prevent trash entries.
Before, the excluded date was simply inserted into calendar2_excluded_dates.
This could be bad. Take this example:
An event has two occurrences. One is deleted, then the other. now we have one
entry in calendar2 and two in calendar2_excluded_dates (plus those in
calendar2_user_relation) that represent nothing.

The new version does this smarter. On each call, it compacts the series as much
as possible and cleans up obsolete entries in the database.