Added input validation to Calendar2_Controller_IndexController
All data used that comes from the client is checked through Cleaner::validate.

For timestamps, we needed a workaround because the client omits the seconds.
So, if the timestamp matches /\d{4}-\d{2}-\d{2} \d{2}:\d{2}/ (the general form
of a timestamp excluding the seconds), we assume it is valid.
If this assumption is wrong, 'new Datetime(...)' should throw an Exception.