Added a 'recurrence' pseudo-attribute to the Calendar2 Model.
This attribute returns a human-readable form of the recurrence rule.

This can be used to intelligently structure the list view of events.
It is impossible to list all single occurrences (remeber infinite events), so we
just use the default jsonListAction but add a value to the model that explains
the recurrence.