Fixed a bug where events from fetchAllForPeriod had no participants.
The cause of this was that Calendar2_Models_Calendar2->copy() did not find() the
occurrence but just copy the data array. This caused ->_storedId to be empty,
which made ->_fetchParticipants think that the object has no database row and
thus not query for participants.

Also, the original start date was not copied.