Added tag handling to Calendar2.
Because we don't have a 1-to-1 relation between logical units (events) and
database entries, tag handling does not work out of the box.

We need to prevent the client from deleting the tags on a jsonDeleteAction call,
because this could just mean that a occurrence will get excluded, and do it on
the server if that's not the case.

We also need to copy the tags when saving to a new row.
Also, when the client updates a event, it needs to use the new id instead of the
old one to save the tags, as the id could have changed because of a splitted or
extracted event.