Fixed PHPROJEKT-51: Adding recurrence to non-recurring events produces meaningless popup.
This was solved by adding a field '_originalRrule' to phpr.Calendar.Form and
overriding one of the initialization Functions to store this after loading the
data from the server.

When preparing the submission, we now check if this change has added a
recurrence to the event and set _multipleEvents for the user instead of showing
a popup.