Make the upgrade screen be friendly to non-upgrade users and prevent them from upgrading.
The standard check for non-admin users in Core_IndexController has been disabled
for the upgrade controller. It sent regular users into a redirect loop
(login->upgrade->login...). We now show a screen explaining the situation
instead.

Because we now can't rely on Core_Indexcontroller::init keeping unauthorized
users out, we have to check for this ourselves.