Show meaningful warnings for participants in month view.
Before, we have shown 'some participants have {not yet accepted,rejected} your
invitation' to participants, too. This is pretty meaningless.

Now we check if the current user is the event's owner, and if that's not the
case we only show a 'you still need to respond' message if the status is still
pending. (If the user rejected, he knows he did. No need to show that.)