The default {start,end}date for new todos is now that of it's project.
I overwrote the jsonDetailAction function in the index controller. If it's a new
record ($id is empty) then we adjust the start and end date, else we call
parent::jsonDetailAction.

Some code in the client that overwrote the start date was removed. It is now
always set on the server.