Fall back to english before returning the untranslated message on the server.
Translations are sometimes used to abbreviate long strings in the code.
If this is the case and the user is using a language that has no translation for
a given (abbreviated) string, the user should not see the abbreviation but the
english version of that string.

This is implemented as follows:
If a Phprojekt_LanguageAdapter has a non-'en' locale, it creates a
_fallbackAdapter with locale 'en'. If then the translation for a given message
was not found, the translation returned by the fallback adapter is used instead.

If the locale is 'en' and a translation is not found, the original message is
returned (as before).