Fall back to english on the client, too.
Unfortunately, this is not as simple as the server side part.

The client does not know anything about locales, it just get's a two-step hash
with modules/messages as keys and translation as values.

To get around that without rebuilding that part, the server now includes a
element "_fallback" in the array with the modules as keys if the selected locale
is not 'en'. The value of this key is again module/message => translation, but
for english.

The client checks for a _fallback element if the translation could not be found
and searches there if it exists.