Somewhat fixed this with normal Forms (not submodules)
As said before, we need to connect the submitForm method to the onSubmit event
of the form instead of the onClick event of the button.
Sounds easy, right?

We actually have a variable number of <form> elements, even without submodules.
Each tab has it's own form element. This is probably also the reason why submit
buttons on every tab are so hard to implement.

This commit simply connects the onSubmit event of all forms in the #detailsBox
with the submitForm method.
This works fine as long as you just open the form, edit stuff and save.
Submodules are not that easy. They might open other forms that need specific
onSubmit handlers.