Sanitize every input field by default as html/string
It is possible to introduce XSS on varchar fields. Therefore we try to
sanitize every field that is not otherwise santized as html/string.

The proper solution is to sanitize on the client, but for 6.0.x a work around
is sanitizing everything on server side instead of santizing parts on the client
and parts on client side.

Signed-off-by: David Soria Parra <david.soria_parra@mayflower.de>