Overwrite jsonSaveMultipleAction in Calendar2_Controller_IndexController
This method is used in the drag and drop implementation. We need this customized
because we need both the id and the recurrenceId of the event to change.

I also added a helper method to Calendar2_Models_Calendar2 that allows to find
by id and recurrenceId instead of id and date of occurrence.