Fixed bug that events that start before the selected period not being shown.
The check whether an event lies in the selected period was just against the
start date of the event. This prevented events that started before and ended in
or after the selection period from being returned by fetchAllForPeriod.

The fix is that we give the rruleHelper a DateInterval $duration, which he then
adds when checking.