Fixed the way the recurrence id is passed through the client.
The way the recurrence ID was passed through was not good. We can't use the
start value because it will take precedence over the value sent from the server.
While this is usually the same (which is why this hasn't been a problem yet) it
doesn't work if we pass 20100530T022800-style recurrence ids.

calling this.setUrl(arguments) instead of this.setUrl() and using undeclared
parameters is a horrible way to do these things.
In the light of newClient, I don't want to put the amount of energy into this
that a clean solution would need.