Save the complete original data in the form, not only the rrule
This is needed because we have to specify which occurrence of the event we want
to change on the server.
This was originally done by adding a recurrenceId value to the events which held
the original start datetime of the event.

Because we don't want this to be confused with caldav recurrenceIds, and because
it isn't neccessary to duplicate this info, this has been changed.