Remove most of the IE8 memleak
I added the pagemanager to be able to track page teardown and loading.
I also added the garbageCollector who can be used to keep track of all the
connected events and created dijit widgets.
I changed the BasicData form to only render the wysiwyg editor if it is needed
because it uses iframes and iframes are not collected in IE8, so they are a
constant memleak.
I replaced the way the page is loaded and replaced most of the location.hash
code to be able to integrade the pagemanager.
We are now able to save arbitrary data into the hash and we no longer depend
on the location.hash to change pages or create certain events.