Don't exclude modified events in caldav.
The problem is that in caldav, if a single occurrence of a recurring event is
excluded and has a specific calendar object for that same occurrence, the event
is not shown. This is exactly what we do when you modify an event out of
a series in P6.

To make this work, we filter the excluded dates for overwritten ones don't send
those to the client.