integrate viewmanager into Core Main.js
replace all dojo.byId calls that do not work anymore because of the viewManager.
create a contentMixin to deal with the specific core module needs for example that the details box has to be in a
brodercontainer to be styled correctly.
also we use the templatewrapper for templating instead of the render function.