Don't hang when an unknown module is saved as last location
If the user has an unknown module as url hash or cookie, the last thing the
initialization routine would do is to publish ${MODULE}.load(), which is unheard
if ${MODULE} doesn't exist.
This problem occurs after upgrading. If a user was at a module which is removed,
the only way to get out is to clear all cookies, which is extremely annoying.

This fix just inserts a list of all known modules and sets the module to
"Project" if the saved module is unknown.