include noAction options in pagemanager
the new options allows the pagemanager to only do some cleanup work and
put the new state into the hashtag.
no actions will be taken concerning the active modules.

We need this flag to be able to remove the "id" element from the current
state without triggering a module reload.

The long time goal here is to get rid of the current pagemanager api as it
is legacy ridden because of the old comma separated format.