Fix issue with displaying administration/settings
the administration and settings view did not work
properly with the new dialogform because the flow
was basically based on luck.
i restructured it a bit to work again, but the
primary goal should be to delete all occurences
of phpr.(module/parentmodule/projectId/etc.)
also the core module is full of special cases.