Use error_log($message) instead of die($message) during startup.
We can't use our own error log, because we need the configuration to find it's
location. die() just prints to standard output which finally gets sent to the
client.

This might send things to the client that it should not see, and it prevents
administrators from examining the error after it happened.

By using error_log($message); die(); we print it to the webserver's error log,
which doesn't have these problems.