Refactor form rendering to use widgets instead of html
the new rendering gives us true javascript object.
so we can easily attach javascript functions and don't have to work
with inline javascript in inline html in javascript.
also Field.js now has its own garbage collection which is triggered if it is
destroyed.
so the parent only have to destory the field generator to collect all used
widgets.