In saveSingleEvent, call save() if the event has no occurrence (again).
Originally, this was done by just checking if $this->rrule is set. This didn't
work because some code set rrule to '' and then called saveSingleEvent, which
then caused all future events to be updated instead of just this single
instance.

The way it was until now (not calling $this->save()) didn't work because the
code in saveSingleEvent relies on the event being part of a series, which should
be okay.
As _originalData was added to ActiveRecord, we can use that instead,
which should solve both problem.

Trying to write clean code, I implemented the check in a function called
hasMultipleOccurrences.