Fix projects not being saved.
The problem was that a refactoring of the save helper removed the call to
setParentNode, which conditionally calls save() on the real activerecords.

To force this call, the pre-refactoring code set $node->projectId = 0, which is
what this commit does, too. Maybe someone should look over the tree
implementation and it's coupling to activerecord/item.