Allow users all rights on things they created.
This is implemented by checking if an item is new and returning true if it is in
Phprojekt_Item_Abstract->hasRight().

If the item is new, $this->id is null, too. This makes us call
Phprojekt_Right::getRightsForItems($moduleId, $projectId, $userId, array(null))
later, which fails with an sql error (something about "WHERE xxx IN ()")

This caused a bug which didn't allow non-admin users to add new contacts