fix bug in module designer with special tab names
the moduledesigner used eval to fetch the tabs it created.
this leaded to errors when the tab names contained for examples arithmetic
operators like "-".
We now use a global hashMap to store the tabs, which is a little bit better but should
be refactored at some point in time because it causes memleaks like the old solution.