restructure form rendering
we don't concatenate html anymore, instead we use the template wrappers.
this way we can attach all the event handlers before we add the form to the dom.
this allows us to postpone the dom insertion of the form to the very end which
results in much faster form rendering because we don't trigger dom reflows all
the time.
also we can redure our dependence on dom id's in favor of dojo attach points.