translate "Basic Data" in the module tabs
Previsously, we modified the data in the datastore and used that modified data
to check whether we already added a basicData tab.
we now just clone it, so we don't modify our data in the datastore.
As a nice side effect, we can translate "Basic Data" in the tab label.