rename timecard booking store and use the merged favorites in timecard main.js
as these are kind of interwoven right now and i can't commit single character
changes, these two go in one commit.
What it does:
it renames the timecard booking store to phpr.timecard.Store and changes all
references from this._bookingStore to this._store.
Also it uses the new getMergedFavoriteProjects function to get the project list.