Only check the server requirements on the first page of the setup.
We checked the requirements on every request of the setup until now.

As we are going to write the config file in the middle of the setup, keeping
this behaviour up would result in errors on every page after writing
configuration.php, as it's absence is checked with the server requirements.