fix rendering of submodules
the submodule rendering did not use the delayed rendering process like the
normal form.
we now just pass in the outer form and use the outer forms addTab method to
add the submodule.
This still has some issues like dynamic sizing but it will be much more work
to fix these.