save: only set rights if we send new rights
This fixes PHPROJEKT-353.
In case no 'dataAccess' field was specified we just ignore it and use
the field from the database. This allows us to edit fields without providing
'dataAccess'.

The inline editing of PHProjekt uses the jsonSaveMultiple call that sets
a column without providing dataAccess, causing the save method to abort
as no rights were set. With this commit the old value will be used and
not overwritten.