fix multiplefilteringselect and settings handling
the settings were handled wrongly.
the problem was:
We could provide a custom function to unserialize a settings value, but we could
not specify a function to fetch a custom settings value, for example if it
resides in its own table (like proxies).
Now we check for the existance of a get<setting>Value function and use it
if it is present.