provide a global userStore which is used as the default userStore.
also provide the possibility to use a local userStore in the form by setting
this.userStore.
this was needed because for example the administration modules did not have
a userList and where unable to call some of its inherited functions because of
that.
we now just use the global one which is more efficient anyways.