Don't filter out versions above the current one when parsing db files.
I can only imagine that the idea behind this was the ability to set the database
state by setting the project version. We have real migrations, so this can't
work.

The filtering made us use 6.x.y-dev or 6.x.y-RC because we are setting the
version to 6.x.y-dev when developing. Changing it to 6.x.y when releasing is
an unneccessary risk and was not done.

After this commit, we can just use 6.x.y when we are on 6.x.y-dev.