Make caldav usernames case insensitive
They were "partially" case insensitive. We rely on two authentications. One is
our standard phprojekt authentication (which gets the httpauth values), which
was case insensitive. But the Sabre ACL takes the username directly from the
http headers and is case sensitive.

To work around this, we set $_SERVER['PHP_AUTH_USER'] to the real username after
successful authentication with phprojekt.