create a replacement for the EditorContainer Dialog mess
We had no abstraction for the html editor in a dialog.
That was messy because we just glued it together in a template.
We now introduce a real wrapper widget.

This one has the advantage that it is actually working, the old one was somehow
broken an missing many elements and functionality.
Also it is much smaller than the EditorContainer we previously used.