fix menubar styling
i had to reorganize the menubar to fix some issues with the layout:
as we have now buttons left and right, drawing left borders on all td's is not
working anymore, instead we use left/right classes to determine where it is.
furthermore the logout button worked only when clicked directly on the icon,
which was confusing as the whole td created a pointer cursor.
modifying the dijit button to stretch to the whole size broke the layout, so
i just reimplemented it as a clickable td with a icon in the middle.