Delete unecessary modules and delete them from database
We remove the unecessary code that we don't need for PTC anymore.
Calendar2 depends on WebDAV and we need the Calendar2 API, therefore
both stay. The migration do delete the modules from the module folder
is done in Timecard. It should be done in Core, but as Core doesn't have
a "version" the migration will fail. Therefore we do it in Timecard
as this is part of the new timecard anyway.