Fix formatting in firefox
Firefox shows two lines, even if the field is empty. I assume it is
a placeholder for scroll bars that are never shown (Firefox wraps lines, even
without word separators) and setting overflow: none; fixes it.

It has to be !important because otherwise dojo overwrites it with
overflow-x: auto; in the element-style.