Throw an exception in case the user doesn't have an associated contract
In case the user doesn't have a contract we need to find a way to make
sure he knows about this and can fix it. Therefore we cannot just
return 0 as this is not a special case. There might be contracts (e.g.
freelancers) that will have a 0h contract.