Encapsulate the temporary variables in a function instead of a method object.
With javascripts anonoymous functions and closures, we don't need method
objects. Instead, we can just declare the function and define it in an anonymous
function, that encapsulates the temporary variables.

This has the advantage that the code that uses the method is not aware of the
way it is built.