make the monthgraph/monthtable load depend on the projectchooser change event.
we don't load the monthGraph right away, because it will be reloaded anyway
once the projectChooser starts up and fires it's 'change' event.
By just settings the monthState, the projectChoosers change event will
trigger a graph/table reload once it has finished starting up.
This prevents a double initialization of the monthGraph which was sometimes
visible as a visual 'flickering' of the graph