fix startup sequence of the javascript
main pulled in Api.js which immediately send off a request to cache the project
names. at this point in time, no csrf token was set and the first request failed.

we now cache the data on the first project fetch and replace the old function
with a cachung wrapper.