Use today as default endday
The current calculations for the current month use the whole month it's
calculation period. This results in a "hours to work" to be calculated
for the full month. In fact the user is interested how much he has to
work till today. So we change the default end date to today.