Use complete month range when generating statistics
Per default our range for the timecard calls is from the first of the month
to the current date. This works fine for calculating the overtime, etc. To
correctly generate the statistics for the whole month we need to get all
days of the current month. Therefore we define the end date manually.